[Error] Move `seek_location` from parser to error files
Not sure why it was in the parser in the first place... either way,
this allows `error.c` to no longer rely on parser.h, which is a good
dependency to break if you ask me (especially when attempting to reuse
this code for other projects)
