+++ src/parser.h
@@ -18,13 +18,4 @@ NODISCARD AST *parse(span source, const char* filename);
 /// Get the string representation of a token type.
 NODISCARD const char *token_type_to_string(enum TokenType type);
 
-/// Seek to a source location.
-void seek_location(
-  span source,
-  loc location,
-  u32 *line,
-  u32 *line_start,
-  u32 *line_end
-);
-
 #endif /* COMPILER_PARSER_H */
